fix: #186 webContextExtractor docs & optimize by  domain tail extract
